create directory automodsumm and move core.py into it
