+++ docs/plasmapy_sphinx/autodoc/automodapi.py
@@ -324,7 +324,7 @@ from sphinx.locale import __
 from sphinx.util import logging
 from typing import Any, Callable, Dict, List, Optional, Union
 
-from ..core import AutomodsummOptions, option_str_list
+from ..automodsumm.core import AutomodsummOptions, option_str_list
 from ..utils import default_grouping_info
 
 if sys.version_info >= (3, 0):
@@ -802,7 +802,7 @@ class ModAPIDocumenter(ModuleDocumenter):
 def setup(app: Sphinx):
     """Sphinx ``setup()`` function for the :rst:dir:`automodapi` functionality."""
 
-    from ..core import setup as setup_automodsumm
+    from ..automodsumm.core import setup as setup_automodsumm
 
     rtn = setup_automodsumm(app)
